Replace extract distribution example with distribution to json.
* Examples were nearly identical.
* Unflatten population-wide number features.
